* pmsubsys.py (pmsubsys): Rename metric_name to metric_pmids,
  metric_desc to metric_descs, metric_value to metric_values. Remove
  is_implemented.
* pmatop.py: Initial version.

+++ src/libpcp/src/err.c
@@ -211,7 +211,7 @@ pmErrStr_r(int code, char *buf, int buflen)
 	    if (sp != NULL) {
 		sp++;
 		for (p = sp; *p != '\0'; p++) {
-		    if (!isdigit(*p)) break;
+		    if (!isdigit((int)*p)) break;
 		}
 
 		if (*p == '\0') {
